llc: Fix comparison warning
Fixes: Jenkins build #609 warning

Addresses:
  llc.cpp:56, GNU C Compiler 3 (gcc), Priority: Normal
  comparison between signed and unsigned integer expressions

Sponsored-by: On-Waves ehf

+++ src/llc.cpp
@@ -49,7 +49,7 @@ void gprs_llc::put_dummy_frame(size_t req_len)
 	static const uint8_t llc_dummy_command[] = {
 		0x43, 0xc0, 0x01, 0x2b, 0x2b, 0x2b
 	};
-	static const int max_dummy_command_len = 79;
+	static const size_t max_dummy_command_len = 79;
 
 	put_frame(llc_dummy_command, sizeof(llc_dummy_command));
